assignment-4/requirements.txt
numpy>=1.23.0
pandas>=1.3.3
scikit-learn>=0.24.2
matplotlib>=3.4.3
seaborn>=0.11.2
scipy>=1.6.0
jupyter==1.0.0
yellowbrick>=1.5
mdptoolbox-hiive==4.0.3.1

gym~=0.26.2